194. Handling Command Clicks(Section 13: Navigation From Scratch)
All right, so last thing we need to do is make sure that on Mako is holding down command and clicking

a link were on Windows, holding down control and clicking on a weak link, opens that link inside of

a new tab.

So to do so, we're gonna go back to our link component's.

We're going to find the unclick function.

And then inside of here, we're going to add in a little check on that event.

We're going to say if E bent dot Medda key or event dot C TRL key.

These are both boolean properties on Mac OS.

We make use of metal key on windows.

We make use of control key.

These are going to be either true or false to indicate whether or not that respective key was held down

when a user clicked on this thing.

So if either these are true, then we're going to not want to run any of this stuff.

Instead, we're going to want to allow the browser to just do its normal thing, which is to open up

a new tab and navigate to the ATF on this link.

So if either of these cases are true, we're going to return early.

And that's pretty much it.

So we can now save this.

Go back over and do a quick test.

So now on Windows, you can hold down control and click on the link and on Mac OS, you can hold down

command and click on the link.

I can do that right now.

And you'll see that it does, in fact, open up a new tab navigated to the appropriate.

You are El.

All right.

Well, that is pretty much it.

That is some basic navigation without making use of react router or anything like that.

Now, remember, we are going to cover react rather later on inside the course.

But I think you're going to see that a lot of stuff we just put together looks very familiar inside

of reactor router.

You're going to have a link component that works very similar to the one we just put together.

And in React router, you are also going to have a root component that works very similar to the one

we just put together as well.

So even though we didn't make use of react router, believe it or not, you actually have some knowledge

around how it works.

All right, well, this has been a lot of fun.

Let's take a pause right here and continue in just a moment.