196.2. Refactoring the SearchBar(Section 14: Hooks in Practice)
We might decide to delete it and instead write that function in line or the on change handler of the

input element.

So if we wanted to do that, we'd put inside of your event.

And then set term event target value.

That is not strictly required, but it does help clean up our function component just a little bit and

make it easier to read.

So if you make that change again, you no longer need the on term change or on input change, whatever

we had called it, event handler to cleared up your separately.

All right, so now the last thing we're going to do is go down to the class component that no longer

has any code inside of it.

We can delete that thing entirely because we have completed our refactor.

Let's save this.

Go back over to our browser and we're just gonna make sure that everything works as expected still.

So I'll do a search again for Forest.

Yep, surge worked.

I can still select a video and play it.

Yes, that's one refactored down.

Let's take a pause right here and then start to refactor our app component in just a moment as well.