198.1. Removing a Callback(Section 14: Hooks in Practice)
They're going to try to, well, just remove some variables if we can.

We're going to try to extract a couple of variables in two separate helper functions and so on.

The first change we're gonna make around our app component is going to very simple and straightforward

in nature.

I want you to take a look at the on video select function right here.

This is a one line function.

And anytime we have a one line function callback, usually a good sign that we should probably just

try to provide that function in line to whatever prop it is going off to.

In this case, on video, select is being provided to the video list component rather than making a

separate on video select function.

We can probably just define that function or provide it in line right here instead.

And that hopefully will make our overall function component just a little bit easier to read and understand.

I'm going to find video list going to temporarily give myself a little bit of space here.

Just you can read this code very easily.

Rather than providing a separate function, I'm going to provide an in-line function, though.

Recall that on video Select is going to be given a video that the user just clicked on.

And then inside there, we're going to update our set selected video or just selected video piece of

state.

So we will call set selected video and provide that video.

So we can now remove the original function right here.

Okay, that looks good.