198.2 Removing a Callback(Section 14: Hooks in Practice)
Now, when we look at this code right here, I want you to notice the pattern that we have.

We are receiving some list of arguments in this case, just one argument.

We then turn around and provide that exact list of arguments.

Again, in this case, just one argument to the set selected video function.

Anytime you see some code like this where you take one argument and just pass it directly through in

the same order to some other function, that is a sign that we could do a little refactor so we can

technically take this entire function, remove it and just replace it with set selected video.

Like, so this is a hundred percent equivalent to the code we just had a moment ago.

Well, we had a moment ago.

We're just taking one function or one argument and passing it directly through the what is the purpose

of making that erro function?

There is no purpose to it.

We can just provide that selected video.

And it's 100 percent equivalent.

We can out save this and then go back over and test it inside of our browser.

The back over here.

I still see everything up here.

I can click on a video.

It gets selected and I can play it.

All right.

Well, I'd say that looks pretty good.

OK.

That's one little change.

Now, obviously not a big change, but whatever, it's enough.

Let's take a pause here.

Welcome back.

Next video we're going to discuss making a much larger change or much larger refactor to our app components.