201.2 Extracting Video Logic(Section 14: Hooks in Practice)
So if we want to, we can follow the same convention as the you state hook or alternatively we can do

what is a little bit more common in the JavaScript community.

And that would be to return an object.

That has properties of videos and on turn cement.

This is totally up to you.