201.3 Extracting Video Logic(Section 14: Hooks in Practice)
There's no requirements in how you return values from a custom hook.

You can either follow the convention of react, which is to return a raise that has stuff inside them.

Or you can do the more common JavaScript convention, which is to return an object.

In this case, I'm going to follow the REAC Convention, which is to return an array.

Now, as I look at this code, the other thing I want to mention very quickly is that the name of this

function on term simit, not the best name in the world on term Simit, made a lot of sense inside of

our abdon jazz component because it was a callback that we wanted to run.

Anytime a term was submitted.

But now that we're looking at a custom hook.

Well, I don't know on term simit, not quite as useful a function name.

The one very small improvement we might make here would be to rename this function to simply search

is that's what this function really does.

It searches for a list of videos.

OK.

So believe it or not, that is our custom hook up.

One last thing we do need to update that unturn summit to search as well.

So this now is our custom hook.

We can make use of use videos in any component.

And as long as we provide a default search term, we will always get back a list of videos and a function

that we can use to update that list of videos as well.

We've built a custom home.

We'll take a very brief pause and then wired this thing back up inside of our APT JS file in just a

moment.