202.2. Using the Custom Hook(Section 14: Hooks in Practice)
Right now, we're going to do what REAC says and put in default search term like so.

This actually does have a good benefit to it.

Now, if we ever decide to call use videos again, but provide a different default search term instead

of buildings right here, maybe we change it for some reason based on some outside prop that flows into

the up component.

Then we're going to make sure that we automatically refresh our list of default videos.

Again.

So we save that.

Now, that warning should go away.

I do still see a lot of other warnings and errors.

But all these are truly fine, not really related to the record that we have written.

OK.

Well, that is pretty much it.

So, again, custom hooks.

Very tricky to understand.

The first time you see it, you might be thinking.

All right.

So many questions.

What's going on here?

What's going on here?

Like I said, this is one of the things where you just have to write your own code.

You have to do this on your own a little bit before it can make any sense.

And there's really not any tutorial or magic video out there that's going to give you a perfect understanding

of how custom hooks work.

Well, having said that, let's take a pause here and continue in just a moment.