214.2 Testing Our Example(Section 15: Deploying a React App)
Let's try doing a couple of quick examples here.

I'm going to refactor this store of dispatch and instead just pass the result of Culin create policy

directly to it.

So I got to put create policy in here and I'll enter in Alex and 20 and now I do not need to create

the action ahead of time.

OK.

So that's going to create a new policy for Alex.

I'll now try to create a new policy for Jim and let's say Jim has to pay $30 and maybe how about I don't

know.

Well say Bob has to pay $40.

So we've now created three policies and we've paid in a total of 20 30 and 40 to our company.

So 20 plus 30 plus 40 plus the default of 100 that our company initially had is 190.

So we see the results over here we have a counting of 190 and then we have policies created for Alex

Jim and Bob.

Now if we wanted to we could have one of our different people tried to create a claim so I could say

Stuart dispatch create claim and we'll say that Alex wants to create a claim for about $120.

So now when we do our get States we'll see that we've now deducted $120 smart company.

So we're now down to $70.

Alex still has a policy as he did before and now we can see that we also have a claim history object

here to reflect the fact that Alex just tried to withdraw a hundred dollars.

The number is right there are $120 excuse me and the name of the person who withdrew that was Alex.

So we can say that maybe Jim gets in a accidence well so it could do a story about dispatch create claim

for Jim and maybe Jim Saxton it wasn't so bad.

So he only needs about $50.

And so now we'll know.

So we have $50 less than we had before and we now have two different claims inside there.

And then finally let's test out the last action we'll do our last action creator which was policy

so I'll do a delete policy and pass in Bob because Bob realized he never made a claim and he doesn't

really need any insurance at all.

And so now when we see the results of our console log we see that only Alex and Jim now have a policy

and we still have all the same claim history and accounting stuff that we had before.

OK.

So that's it.

That is redux in action.

Now let's take a pause right here.

When we come back the next video I'm going to point out some very distinct things that I want you to

notice about the code that we just wrote right here and this is going to help you understand why we

use redux inside reactor application and why it's so widely used in industry.

So quick pause and I'll see you in just a minute.