215.1 Important Redux Notes(Section 15: Deploying a React App)
In the last section we finally finished writing code to model our insurance company.

So in this video I want to do a wrap up on some of the big things that we learned to YOUR about redux

at this point.

I don't really expect you to memorize a lot of the intricate details here.

I just wanted to have an overall idea of how the redux library worked.

So let's get to it.

All right so first things first we're going to start to really learn this redux cycle this redux cycle

that you see right here has all the same boxes inside of it as the diagram that we're looking at just

a second ago.

So you can now remember from our analogy the purpose of the action creator that's like a person dropping

off for the purpose of our action purpose of dispatch and so on.

So with the redux cycle we essentially start off at the very top up here.

Any time we want to change the state or the data of our application we're going to call an action creator

calling an action creator is going to produce an action object this action object describes exactly

how we want to change data inside of application that action object gets fed to the dispatch function

which in turn is going to make copies.

We can imagine of the action object and feed those copies to each of our different reducers in turn.

The producers are going to run.

They're going to process those actions are the forms that came in modify their data and then eventually

return some new data that data that gets returned gets forint into some new state object.

So then we wait until we need to somehow updates our state again at some point in the future.

So now we have a somewhat reasonable idea of the purpose of each of these steps inside this diagram.

I want to give you some more insight about what's going on inside of redux now.

So you can flip back over to my code over here.

All right so the first thing I want tell you about was the purpose of this combine it reducers call.

Remember the different readers we put together were simple functions.

All these different functions need to be somehow wired up together.

We wire up all the different reducers together by making use of the combine reducers function.

You'll notice that we have different keys on here of accounting claims history and policies because

we used these very specific key names.

We ended up with a state object.

Remember that's what's being called a logged right here.

That's the result of the con. log of stored get state.

We ended up with properties on our state object equal to those key names so we can absolutely feed keys

into this object right here with names different from our reducer.

For example if I wanted to I can change this key of accounting to money we have and now in this example

reruns you'll see that my state object now has a different key for the accounting reducer and now says

money we have is 20.

We do not always have to name these reducers exactly after the Keys but usually by convention we're

going to have them be very similar.